Support refresh to reload with Mr
[Mr][] and [Mop][] serve a similar function to [Browserify][] but
supports refresh-to-reload during development using Mr and whole package
application transformation with Mop. Thus a build step is not necessary
during development, and transforming packages for use in production does
not require any configuration changes in source. These loaders support
Node.js modules and packages with very high fidelity, inferring where
npm installed packages, but with one major caveat, that they cannot
infer the existence of an `index.js` file during development. To avoid
this gotcha, this change adds a `main` property to `package.json`
pointing to the root-level `index.js`, and alters usage of the other
index modules by explicating the `index` at the end of the module
identifier.

[Mr]: https://github.com/montagejs/mr
[Mop]: https://github.com/montagejs/mop
[Browserify]: https://github.com/substack/node-browserify