include all source files in the report
hidden behind the --include-all-sources flag

After the coverage report has been built, manually instrument the files
which have not been visited already, and add those to the report.

When instrumenting the code manually, reset the count in the coverState
object. FunctionDeclaration is given a visit number of 1 when
instrumenting to compensate for function hoisting. We don't want that
in this case, as the file was not actually loaded.