cli: Don't throw strings when there are errors
It makes for confusing output that tells the user something is wrong
with the program, directing them to dig into istanbul's guts to 'fix the
problem', because the file and line number are printed there.

Instead, if we get a string, just output it to stderr, and exit with
status code 1.

If an actual (non-string) error is raised, still throw it, since that
does genuinely indicate that something is probably wrong.

Fix #518