Handle errors when instrumenting 'all sources'
Addresses an issue where the coverage report would silently fail if it
attempted to instrument a file that was 'invalid' for some reason.

Whilst that error once determined was easy enough to rectify (exclusion
rules.) The fact that it silently exited meant it was hard to determine
the cause.

This change ignores any erroring files, but logs the file out to the console.