Make double quote usage consistent in HTML header
Currently, the `fraction` class attribute was used with single quotes, while the others use double quotes. I've converted it to double quotes, to at least have consistency between the header spans. Outside of this, there is still some inconcistency within the templates on attribute quoting, but I considered that to be a too intrusive change.